refactor: 리뷰 등록 시 자동으로 입력 폼이 닫히도록 개선
[#490]
