+++ frontend/src/components/ui/StationDetailsWindow/reviews/previews/UserRatingsSkeleton.tsx
@@ -4,9 +4,9 @@ import Skeleton from '@common/Skeleton';
 
 const UserRatingsSkeleton = () => {
   return (
-    <Box px={4} mb={5}>
-      <FlexBox justifyContent="between">
-        <Skeleton width="7rem" height="2rem" />
+    <Box mb={25}>
+      <FlexBox justifyContent="between" alignItems="center">
+        <Skeleton width="7rem" height="2.2rem" />
         <Skeleton width="5rem" height="2rem" />
       </FlexBox>
     </Box>
