+++ frontend/src/components/ui/StationDetailsWindow/reviews/reviews/ReviewCardSkeleton.tsx
@@ -4,11 +4,10 @@ import Skeleton from '@common/Skeleton';
 const ReviewCardSkeleton = () => {
   return (
     <>
-      <Box p={4} mb={4}>
-        <Skeleton width="10rem" height="1.2rem" mb={2} />
-        <Skeleton width="5rem" height="1.2rem" mb={3} />
-        <Skeleton width="100%" height="1.2rem" mb={2} />
-        <Skeleton width="100%" height="1.2rem" mb={5} />
+      <Box p={2} mb={4}>
+        <Skeleton width="10rem" height="1.2rem" mb={1} />
+        <Skeleton width="5rem" height="1.2rem" mb={2} />
+        <Skeleton width="20rem" height="1.2rem" mb={4} />
         <Skeleton width="5rem" height="1.2rem" />
       </Box>
     </>
